// envs/figma/mock/src/types/ops.ts
export interface TransformTuple {
  x: number;
  y: number;
  w: number;
  h: number;
  rotation: number;
  scaleX: 1 | -1;
  scaleY: 1 | -1;
}

export type TransformMap = Record<string, TransformTuple>;